// src/shared/api/generated/types/NewsArticleDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { NewsCategoryDto } from "./NewsCategoryDto";

export type NewsArticleDto = {
  /**
   * @type number
   */
  id: number;
  /**
   * @type string
   */
  slug: string;
  /**
   * @type string
   */
  title: string;
  /**
   * @type string
   */
  excerpt: string;
  /**
   * @type string
   */
  content: string;
  /**
   * @type string
   */
  coverImageUrl?: string | null;
  /**
   * @type string, date-time
   */
  publishedAt?: string | null;
  category?: NewsCategoryDto | null;
};