// src/shared/api/generated/types/NewsCategoryDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type NewsCategoryDto = {
  /**
   * @type number
   */
  id: number;
  /**
   * @type string
   */
  slug: string;
  /**
   * @type string
   */
  title: string;
  /**
   * @type string
   */
  description?: string | null;
};