// src/shared/api/generated/types/NewsListItemDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { NewsCategoryDto } from "./NewsCategoryDto";

export type NewsListItemDto = {
  /**
   * @type number
   */
  id: number;
  /**
   * @type string
   */
  slug: string;
  /**
   * @type string
   */
  title: string;
  /**
   * @type string
   */
  excerpt: string;
  /**
   * @type string
   */
  coverImageUrl?: string | null;
  /**
   * @type string, date-time
   */
  publishedAt?: string | null;
  category?: NewsCategoryDto | null;
};